requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
attrs==25.4.0
certifi==2026.2.25
cryptography==45.0.7
distro==1.9.0
dj-database-url==2.3.0
Django==6.0.3
django-allauth==65.15.0
edge-tts==7.2.7
frozenlist==1.8.0
groq==1.0.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
idna==3.11
multidict==6.7.1
pillow==12.1.1
propcache==0.4.1
psycopg[binary]==3.2.6
PyJWT==2.10.1
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
python-dotenv==1.2.2
requests==2.32.3
scikit-learn==1.5.2
sniffio==1.3.1
SpeechRecognition==3.14.5
sqlparse==0.5.5
tabulate==0.10.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
whitenoise==6.9.0
yarl==1.23.0
